Remove `domain` prop; move createUniqueCrumbs to PageLayout utils
