--- src/components/breadcrumb/utils.js
@@ -1,10 +1,0 @@
-// cleans breadcrumbs
-// does not add a link if the path or title already exists
-export function createUniqueCrumbs(links) {
-  return links.reduce((arr, item) => {
-    const pathExists = arr.filter((f) => f.path === item.path).length > 0;
-    const titleExists = arr.filter((f) => f.title === item.title).length > 0;
-    if (!pathExists && !titleExists) arr.push(item);
-    return arr;
-  }, []);
-}
